Receiving, updating, deleting, getting api endpoints tested

+++ backend/tests/createTestUsers.js
@@ -1,12 +1,41 @@
 const User = require('../src/api/models/User');
 
 // Asynchronous function to handle user creation with an optional delete
-const createUsers = async () => {
-    console.log('this is called yea?')
+const createUsers = async (del) => {
+    if (del)
+        await User.deleteMany({});
 
-    await User.deleteMany({});
+    await new User({
+        username: 'admin',
+        passwordHash: 'password',  // Assuming these are pre-hashed or will be hashed in a middleware
+        email: 'test@example.com',
+        firstName: 'Giorgos',
+        lastName: 'User',
+        role: 'admin',
+        department: 1
+    }).save();
 
-    const user1 = await User.create({
+    await new User({
+        username: 'manager',
+        passwordHash: 'password',  // Assuming these are pre-hashed or will be hashed in a middleware
+        email: 'test@example.com',
+        firstName: 'Makhs',
+        lastName: 'User',
+        role: 'manager',
+        department: 1
+    }).save();
+
+    await new User({
+        username: 'scanner',
+        passwordHash: 'password',
+        email: 'test@example.com',
+        firstName: 'Test',
+        lastName: 'User',
+        role: 'scanner',
+        department: 1
+    }).save();
+
+    await new User({
         username: 'giorgos',
         passwordHash: 'giorgos123',  // Assuming these are pre-hashed or will be hashed in a middleware
         email: 'test@example.com',
@@ -14,9 +43,9 @@ const createUsers = async () => {
         lastName: 'User',
         role: 'scanner',
         department: 1
-    });
+    }).save();
 
-    const user2 = await User.create({
+    await new User({
         username: 'makhs',
         passwordHash: 'makhs123',  // Assuming these are pre-hashed or will be hashed in a middleware
         email: 'test@example.com',
@@ -24,10 +53,24 @@ const createUsers = async () => {
         lastName: 'User',
         role: 'scanner',
         department: 1
-    });
+    }).save();
 
-    await user1.save();
-    await user2.save();
+    await new User({
+        username: 'validUser',
+        passwordHash: 'validPass',
+        email: 'test@example.com',
+        firstName: 'Test',
+        lastName: 'User',
+        role: 'scanner',
+        department: 1
+    }).save();
 };
 
-module.exports = createUsers
+module.exports = {
+    withDelete: async () => {
+        await createUsers(true)
+    },
+    noDelete: async () => {
+        await createUsers(false)
+    }
+}
